--- a/Makefile
+++ b/Makefile
@@ -118,4 +118,4 @@ sqs-update-mainnet-state:
 
 # Bench tests pricing
 bench-pricing:
-	go test -bench BenchmarkGetPrices -run BenchmarkGetPrices github.com/osmosis-labs/sqs/tokens/usecase -v
+	go test -bench BenchmarkGetPrices -run BenchmarkGetPrices github.com/osmosis-labs/sqs/tokens/usecase -count=6